requirements.txt
# Core Dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Data Processing & Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0

# Database
sqlalchemy>=2.0.0

# Async & Concurrency
aiohttp>=3.9.0
asyncio-throttle>=1.0.0

# NLP & Text Analysis
scikit-learn>=1.3.0
nltk>=3.8.0
wordcloud>=1.9.0

# Data Validation
pydantic>=2.5.0

# CLI & Progress
rich>=13.7.0
typer>=0.9.0

# Reporting
jinja2>=3.1.0

# Excel Export
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
fake-useragent>=1.4.0
